Już się nie wypierdalaja na etapie kompilowania ale dalej chuj
