+++ src/shaders/default.frag
@@ -42,8 +42,9 @@ void main()
 	float specular = specAmount * specularLight;
 	float specular2 = specAmount2 * specularLight;
 
-	FragColor = mix(texture(texture1, texCoord), texture(texture2, texCoord), 0.55) * (0.5f * lightColor * (diffuse + ambient + specular) + 0.5 * light2Color * (diffuse2 + ambient + specular2));
+	//FragColor = mix(texture(texture1, texCoord), texture(texture2, texCoord), 0.55) * (0.5f * lightColor * (diffuse + ambient + specular) + 0.5 * light2Color * (diffuse2 + ambient + specular2));
 	//FragColor = texture(tex0, texCoord);
 	//* lightColor * (diffuse + ambient + specular);
 	//* (0.5f * lightColor * (diffuse + ambient + specular) + 0.5 * light2Color * (diffuse2 + ambient + specular2))
+	FragColor = vec4(1, 0, 0, 1);
 }
